Removed raw_block in favour of updating Cbor trait, refactored Cid crate
