+++ blockchain/blocks/Cargo.toml
@@ -14,7 +14,6 @@ multihash = "0.9.4"
 derive_builder = "0.9"
 serde = { version = "1.0", features = ["derive"] }
 encoding = { package = "forest_encoding", path = "../../encoding" }
-raw_block = { path = "../raw_block" }
 num-bigint =  { git = "https://github.com/austinabell/num-bigint", rev = "f7084a9ed5a2b08d9bfb67790cb4ce9212193f31" }
 
 [dev-dependencies]
