--- blockchain/raw_block/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "raw_block"
-version = "0.1.0"
-authors = ["ChainSafe Systems <info@chainsafe.io>"]
-edition = "2018"
-
-[dependencies]
-cid = { package = "forest_cid", path = "../../ipld/cid" }
-serde = { version = "1.0", features = ["derive"] }
-encoding = { package = "forest_encoding", path = "../../encoding" }
-multihash = "0.9.4"
