--- blockchain/raw_block/src/lib.rs
@@ -1,6 +1,0 @@
-// Copyright 2020 ChainSafe Systems
-// SPDX-License-Identifier: Apache-2.0, MIT
-
-mod raw_block;
-
-pub use self::raw_block::*;
