+++ blockchain/sync_manager/Cargo.toml
@@ -15,7 +15,6 @@ cid = { package = "forest_cid", path = "../../ipld/cid" }
 blockstore = { package = "ipld_blockstore", path = "../../ipld/blockstore" }
 chain = { path = "../chain" }
 message = { package = "forest_message", path = "../../vm/message" }
-raw_block = { package = "raw_block", path = "../raw_block" }
 multihash = "0.9.4"
 state = { package = "state_tree", path = "../../vm/state_tree/" }
 num-bigint = { git = "https://github.com/austinabell/num-bigint", rev = "f7084a9ed5a2b08d9bfb67790cb4ce9212193f31" }
