+++ encoding/Cargo.toml
@@ -9,3 +9,4 @@ blake2b_simd = "0.5.9"
 serde = { version = "1.0", features = ["derive"] }
 serde_bytes = "0.11.3"
 serde_cbor = { version = "0.11.0", features = ["tags"] }
+cid = { package = "forest_cid", path = "../ipld/cid" }
