+++ ipld/cid/tests/base_cid_tests.rs
@@ -1,11 +1,9 @@
 // Copyright 2020 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use encoding::{from_slice, to_vec, Cbor};
 use forest_cid::{Cid, Codec, Error, Prefix, Version};
 use multihash;
 use multihash::Hash::Blake2b256;
-use serde::Serialize;
 use std::collections::HashMap;
 
 #[test]
@@ -121,36 +119,4 @@ fn test_default() {
         // 4 is Blake2b256 code length (3) + 1, change if default changes
         (Blake2b256.size() + 4) as usize
     );
-}
-
-#[derive(Serialize, Copy, Clone)]
-struct TestCborStruct {
-    name: &'static str,
-}
-impl Cbor for TestCborStruct {}
-
-#[test]
-fn test_cbor_to_cid() {
-    let obj = TestCborStruct { name: "test" };
-
-    let enc = Cid::from_cbor(obj, Blake2b256).unwrap();
-    let bz_enc = Cid::from_bytes(&obj.marshal_cbor().unwrap(), Blake2b256).unwrap();
-    assert_eq!(enc, bz_enc);
-}
-
-#[test]
-fn vector_cid_serialize_round() {
-    let cids = vec![
-        Cid::from_bytes(&[0, 1], Blake2b256).unwrap(),
-        Cid::from_bytes(&[1, 2], Blake2b256).unwrap(),
-        Cid::from_bytes(&[3, 2], Blake2b256).unwrap(),
-    ];
-
-    // Serialize cids with cbor
-    let enc = to_vec(&cids).unwrap();
-
-    // decode cbor bytes to vector again
-    let dec: Vec<Cid> = from_slice(&enc).unwrap();
-
-    assert_eq!(cids, dec);
 }
