+++ vm/message/Cargo.toml
@@ -13,7 +13,6 @@ encoding = { package = "forest_encoding", path = "../../encoding" }
 crypto = { path = "../../crypto" }
 derive_builder = "0.9"
 serde = { version = "1.0", features = ["derive"] }
-raw_block = { path = "../../blockchain/raw_block" }
 
 [dev-dependencies]
 serde_json = "1.0"
