+++ vm/message/src/signed_message.rs
@@ -6,7 +6,6 @@ use address::Address;
 use crypto::{Error as CryptoError, Signature, Signer};
 use encoding::Cbor;
 use num_bigint::BigUint;
-use raw_block::RawBlock;
 use serde::{Deserialize, Serialize};
 use vm::{MethodNum, Serialized, TokenAmount};
 
@@ -68,7 +67,5 @@ impl Message for SignedMessage {
         self.message.required_funds()
     }
 }
-
-impl RawBlock for SignedMessage {}
 
 impl Cbor for SignedMessage {}
