+++ vm/message/src/unsigned_message.rs
@@ -6,7 +6,6 @@ use address::Address;
 use derive_builder::Builder;
 use encoding::{de, ser, Cbor};
 use num_bigint::BigUint;
-use raw_block::RawBlock;
 use serde::Deserialize;
 use vm::{MethodNum, Serialized, TokenAmount};
 
@@ -134,7 +133,5 @@ impl Message for UnsignedMessage {
         total + self.value().0.clone()
     }
 }
-
-impl RawBlock for UnsignedMessage {}
 
 impl Cbor for UnsignedMessage {}
